Move cart and order API logic from Storefront and target the frontend to this (#40)
VP-1075
* move api logic from Storefront project for security and business logic moving to server side.
* target and debug the frontend part to new API.
* fix stylecop warnings after merge

https://github.com/VirtoCommerce/vc-module-javascript-shoppingcart/pull/40

+++ VirtoCommerce.JavaScriptShoppingCart.Model/Model/Cart/Shipment.cs
@@ -68,7 +68,7 @@ namespace VirtoCommerce.JavaScriptShoppingCart.Core.Model.Cart
         /// <summary>
         /// Gets or sets the value of weight.
         /// </summary>
-        public double? Weight { get; set; }
+        public decimal? Weight { get; set; }
 
         /// <summary>
         /// Gets or sets the value of measurement units.
@@ -78,17 +78,17 @@ namespace VirtoCommerce.JavaScriptShoppingCart.Core.Model.Cart
         /// <summary>
         /// Gets or sets the value of height.
         /// </summary>
-        public double? Height { get; set; }
+        public decimal? Height { get; set; }
 
         /// <summary>
         /// Gets or sets the value of length.
         /// </summary>
-        public double? Length { get; set; }
+        public decimal? Length { get; set; }
 
         /// <summary>
         /// Gets or sets the value of width.
         /// </summary>
-        public double? Width { get; set; }
+        public decimal? Width { get; set; }
 
         /// <summary>
         /// Gets or sets the value of shipping price.
@@ -102,24 +102,23 @@ namespace VirtoCommerce.JavaScriptShoppingCart.Core.Model.Cart
         public Money PriceWithTax { get; set; }
 
         /// <summary>
-        /// Gets or sets the value of total shipping price without taxes
+        /// Gets the value of total shipping price without taxes
         /// Price + Fee - DiscountAmount;.
         /// </summary>
         public Money Total { get; set; }
 
         /// <summary>
-        /// Gets or sets the value of total shipping price including taxes
+        /// Gets the value of total shipping price including taxes
         /// Total * TaxPercentRate.
         /// </summary>
         public Money TotalWithTax { get; set; }
 
         /// <summary>
-        /// Gets or sets the value of total shipping discount amount.
+        /// Gets the value of total shipping discount amount.
         /// </summary>
         public Money DiscountAmount { get; set; }
 
         /// <summary>
-        /// Gets or sets the value of discount including taxes
         /// DiscountAmount * TaxPercentRate.
         /// </summary>
         public Money DiscountAmountWithTax { get; set; }
@@ -131,7 +130,6 @@ namespace VirtoCommerce.JavaScriptShoppingCart.Core.Model.Cart
         /// Collection of CartShipmentItem objects.
         /// </value>
         public IList<CartShipmentItem> Items { get; set; }
-
 
         /// <summary>
         /// Gets or sets the value of total shipping tax amount.
